requirements.txt
pytest==3.6.4
PyYAML==3.13